+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Releases
 
+## v5.0.0
+
+- **BREAKING**: Index Redirect functions have new configurations.
+- **BREAKING**: hsts is now enabled by default
+- **BREAKING**: Requires aws privder v3.41+
+- Moved Lambda@Edge functions to CloudFront functions.
+- The index redirect function now created using ``index_redirect``.
+- **NEW FEATURE**: index_redirect_no_extension. Match any URL that ends in /<something-without-a-dot>; append /index.html. Example example.com/foo/bar => example.com/foo/bar/index.html
+
 ## v4.0.1
 
 - Added override variables for default cache behavior cache attributes (eg. default_ttl, max_ttl, min_ttl)